Added calibrate mode
On first starting, the numbers need to be set to be reset to zero before starting to tell time.  Any time there is a power outage, on restart the system goes to calibration mode with the LEDs blinking.  Once you have set the numbers to zero, you push the rotary switch and it starts telling time.